+++ .formatter.exs
@@ -2,5 +2,5 @@
 [
   line_length: 100,
   inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"],
-  locals_without_parens: [assert_files_equal: 2]
+  locals_without_parens: [assert_files_equal: 2, assert_images_equal: 2]
 ]