+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This is the changelog for Image version 0.6.0 released on June 5th, 2022.  For o
 
 * Adds `Image.hamming_distance/2` to compare the similarity of two images or two image hashes.
 
+* Adds `Image.radial_gradient/3` to great a radial gradient image. This could be used to composite over another image to add a vignette.
+
 ## Image 0.5.0
 
 This is the changelog for Image version 0.5.0 released on June 24th, 2022.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/image/tags)